Delete worker-plugin (due origin policy)
