+++ package.json
@@ -70,7 +70,6 @@
     "webpack-cli": "^3.3.11",
     "webpack-dev-server": "^3.10.3",
     "webpack-manifest-plugin": "^2.2.0",
-    "workbox-webpack-plugin": "^4.3.1",
-    "worker-plugin": "^4.0.2"
+    "workbox-webpack-plugin": "^4.3.1"
   }
 }
